refactor(text): split text::* into separate files
